src/RobotArmLocalizationPackage/full_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
attrs=21.2.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
c-ares=1.17.1=h9ed2024_0
ca-certificates=2021.10.26=hecd8cb5_2
certifi=2021.10.8=py38hecd8cb5_0
curl=7.78.0=h7bc2e8c_0
cycler=0.10.0=pypi_0
cython=0.29.21=py38hb1e8313_0
dbus=1.13.18=h18a8e69_0
expat=2.4.1=h23ab428_2
ffmpeg=4.3=h0a44026_0
freetype=2.10.4=ha233b18_0
gettext=0.21.0=h7535e17_0
git=2.32.0=pl5262h6cd52fe_1
glib=2.69.1=hdf23fa2_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
iniconfig=1.1.1=pypi_0
intel-openmp=2021.3.0=hecd8cb5_3375
jpeg=9b=he5867d9_2
kiwisolver=1.3.2=pypi_0
krb5=1.19.2=hcd88c3b_0
labelimg=1.8.5=pypi_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcurl=7.78.0=hb8e4fae_0
libcxx=12.0.0=h2f01273_0
libedit=3.1.20210714=h9ed2024_0
libev=4.33=haf1e3a3_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
libnghttp2=1.41.0=h7580e61_2
libpng=1.6.37=ha441bb4_0
libssh2=1.9.0=ha12b0ac_1
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
libxslt=1.1.34=h83b36ba_0
llvm-openmp=12.0.0=h0dcd299_1
lxml=4.6.1=py38h63b7cb6_0
lz4-c=1.9.3=h23ab428_1
matplotlib=3.4.3=pypi_0
mkl=2021.3.0=hecd8cb5_517
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.1=py38h4ab4a9b_0
mkl_random=1.2.2=py38hb2f4e1b_0
ncurses=6.2=h0a44026_1
nettle=3.7.3=h230ac6f_1
ninja=1.10.2=hf7b0b51_1
numpy=1.21.2=py38h4b4dc7a_0
numpy-base=1.21.2=py38he0bd621_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd9629dc_0
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1l=h9ed2024_0
packaging=21.0=pypi_0
pcre=8.45=h23ab428_0
pcre2=10.35=h1de35cc_1
perl=5.26.2=h4e221da_0
pillow=8.3.2=pypi_0
pip=21.0.1=py38hecd8cb5_0
pluggy=1.0.0=pypi_0
py=1.10.0=pypi_0
pycocotools=2.0.2=pypi_0
pyparsing=2.4.7=pypi_0
pyqt=5.9.2=py38h655552a_2
pyqt5=5.15.4=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.0=pypi_0
pytest=6.2.5=pypi_0
python=3.8.11=h88f2d9e_1
python-dateutil=2.8.2=pypi_0
pytorch=1.9.0=py3.8_0
qt=5.9.7=h468cd18_1
readline=8.1=h9ed2024_0
scipy=1.7.1=py38h88652d9_2
setuptools=52.0.0=py38hecd8cb5_0
sip=4.19.8=py38h0a44026_0
six=1.16.0=pypi_0
sqlite=3.36.0=hce871da_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.2=pypi_0
torch=1.9.0=pypi_0
torchvision=0.10.0=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
